Move Titus request and response data objects into more specific package
They don't model our usual immutable pattern (and by necessity going forward they're going to have to be mutable after initial creation in order to sign the requests for security). They're also not used anywhere else currently. Scoping these down into the launchers package helps keep the deliniation clearer for now.
