+++ genie-web/src/main/java/com/netflix/genie/web/agent/launchers/impl/TitusAgentLauncherImpl.java
@@ -23,9 +23,9 @@ import com.github.benmanes.caffeine.cache.Cache;
 import com.google.common.collect.ImmutableMap;
 import com.netflix.genie.common.internal.util.GenieHostInfo;
 import com.netflix.genie.web.agent.launchers.AgentLauncher;
+import com.netflix.genie.web.agent.launchers.dtos.TitusBatchJobRequest;
+import com.netflix.genie.web.agent.launchers.dtos.TitusBatchJobResponse;
 import com.netflix.genie.web.dtos.ResolvedJob;
-import com.netflix.genie.web.dtos.TitusBatchJobRequest;
-import com.netflix.genie.web.dtos.TitusBatchJobResponse;
 import com.netflix.genie.web.exceptions.checked.AgentLaunchException;
 import com.netflix.genie.web.properties.TitusAgentLauncherProperties;
 import com.netflix.genie.web.util.MetricsUtils;
