+++ genie-web/src/test/groovy/com/netflix/genie/web/agent/launchers/impl/TitusAgentLauncherImplSpec.groovy
@@ -24,9 +24,9 @@ import com.netflix.genie.common.external.dtos.v4.JobEnvironment
 import com.netflix.genie.common.external.dtos.v4.JobMetadata
 import com.netflix.genie.common.external.dtos.v4.JobSpecification
 import com.netflix.genie.common.internal.util.GenieHostInfo
+import com.netflix.genie.web.agent.launchers.dtos.TitusBatchJobRequest
+import com.netflix.genie.web.agent.launchers.dtos.TitusBatchJobResponse
 import com.netflix.genie.web.dtos.ResolvedJob
-import com.netflix.genie.web.dtos.TitusBatchJobRequest
-import com.netflix.genie.web.dtos.TitusBatchJobResponse
 import com.netflix.genie.web.exceptions.checked.AgentLaunchException
 import com.netflix.genie.web.properties.TitusAgentLauncherProperties
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry
